Attached arm pivot and new build.
I attached the arm pivot script, it seems to work but it feels and looks (more then likely due to the way i drew the player and it's animations) a bit strange. There's a new build to try it out. I encourage anyone who thinks they can fix it to do so.